Fix CI: rector/pint findings and peck dictionary gaps
composer test's rector --dry-run step (never run locally before
pushing, only pint/phpstan directly) caught three real findings in
the previous commit: a redundant (int) cast on curl_getinfo() (already
returns int for CURLINFO_HTTP_CODE), a redundant explicit `color: null`
named arg matching the parameter's own default, and a missing arrow
function return type. Applying rector's own fix for the last one left
an inline fully-qualified \Livewire\Component reference that pint's
class_definition rule wants as a proper import instead.

peck (the typo checker) also flagged six legitimate technical terms
introduced by the categories-pivot change ("repoints", "dedup",
"unscoped", "deduplicated", "discoverable", "browsable") that aren't
in its dictionary — added to the project's existing peck.json ignore
list rather than rewording correct English into something clunkier.

+++ app/Services/Curl/Request.php
@@ -81,7 +81,7 @@ class Request
         $response = curl_exec($ch);
         $curlErrno = curl_errno($ch);
         $curlError = curl_error($ch);
-        $httpCode = (int) curl_getinfo($ch, CURLINFO_HTTP_CODE);
+        $httpCode = curl_getinfo($ch, CURLINFO_HTTP_CODE);
         curl_close($ch);
 
         if ($response === false || $curlErrno !== 0) {
